Crate msi

A library for reading/writing Windows Installer (MSI) files.

A Windows Installer file, or MSI file, represents a Windows software package and a declarative description of how it should be installed. An MSI file consists of a relational database stored within a Compound File Binary file.

Structs

Column
A database column.

ColumnBuilder
A factory for configuring a new database column.

Delete
A database query to delete existing rows.

Expr
An expression on database rows that can be used in queries.

Insert
A database query to insert new rows.

Package
An MSI package file, backed by an underlying reader/writer (such as a File or Cursor).

Row
One row from a database table.

Rows
An iterator over the rows in a database table.

Select
A database query to select rows.

StreamReader
An IO reader for an embedded binary stream in a package.

StreamWriter
An IO writer for an embedded binary stream in a package.

Streams
An iterator over the names of the binary streams in a package.

SummaryInfo
Summary information (e.g. title, author) about an MSI package.

Table
A database table.

Tables
An iterator over the database tables in a package.

Update
A database query to update existing rows.

Enums

CodePage
A Windows code page.

ColumnCategory
Indicates the format of a string-typed database column.

ColumnType
A database column data type.

PackageType
The type of MSI package (e.g. installer or patch).

Value
A value from one cell in a database table row.

Functions

open
Opens an existing MSI file at the given path in read-only mode.

open_rw
Opens an existing MSI file at the given path in read-write mode.