Enum msi::CodePage

pub enum CodePage {
    Windows1250,
    Windows1251,
    Windows1252,
    Windows1253,
    Windows1254,
    Windows1255,
    Windows1256,
    Windows1257,
    Windows1258,
    MacintoshRoman,
    MacintoshCyrillic,
    UsAscii,
    Iso88591,
    Iso88592,
    Iso88593,
    Iso88594,
    Iso88595,
    Iso88596,
    Iso88597,
    Iso88598,
    Utf8,
}

A Windows code page.

Code pages are a legacy Windows mechanism for representing character encodings, but are still used within the MSI file format.

Not all Windows code pages are supported by this library yet. See Wikipedia for a complete list of valid code pages.

Variants
Windows1250

Windows-1250 (Latin 2)

Windows1251

Windows-1251 (Cyrillic)

Windows1252

Windows-1252 (Latin 1)

Windows1253

Windows-1253 (Greek)

Windows1254

Windows-1254 (Turkish)

Windows1255

Windows-1255 (Hebrew)

Windows1256

Windows-1256 (Arabic)

Windows1257

Windows-1257 (Baltic)

Windows1258

Windows-1258 (Vietnamese)

MacintoshRoman

Mac OS Roman

MacintoshCyrillic

Macintosh Cyrillic

UsAscii

US-ASCII

Iso88591

ISO-8859-1 (Latin 1)

Iso88592

ISO-8859-2 (Latin 2)

Iso88593

ISO-8859-3 (South European)

Iso88594

ISO-8859-4 (North European)

Iso88595

ISO-8859-5 (Cyrillic)

Iso88596

ISO-8859-6 (Arabic)

Iso88597

ISO-8859-7 (Greek)

Iso88598

ISO-8859-8 (Hebrew)

Utf8

UTF-8

Methods

impl CodePage

fn from_id(id: i32) -> Option<CodePage>

Returns the code page (if any) with the given ID number.

fn id(&self) -> i32

Returns the ID number used within Windows to represent this code page.

fn name(&self) -> &str

Returns a human-readable name for this code page.

fn decode(&self, bytes: &[u8]) -> String

Decodes a byte array into a string, using this code page. Invalid characters will be replaced with a Unicode replacement character (U+FFFD).

fn encode(&self, string: &str) -> Vec<u8>

Encodes a string into a byte array, using this code page. For non-Unicode code pages, any characters that cannot be represented will be replaced with a code-page-specific replacement character (typically '?').