Enum msi::ColumnCategory

pub enum ColumnCategory {
    Text,
    UpperCase,
    LowerCase,
    Integer,
    DoubleInteger,
    Identifier,
    Property,
    Filename,
    WildCardFilename,
    Path,
    Paths,
    AnyPath,
    DefaultDir,
    RegPath,
    Formatted,
    KeyFormatted,
    Template,
    Condition,
    Guid,
    Version,
    Language,
    Binary,
    CustomSource,
    Cabinet,
    Shortcut,
    Url,
}

Indicates the format of a string-typed database column.

Variants
Text

An unrestricted text string.

Examples

assert!(msi::ColumnCategory::Text.validate("Hello, World!"));

UpperCase

A text string containing no lowercase letters.

Examples

assert!(msi::ColumnCategory::UpperCase.validate("HELLO, WORLD!"));
assert!(!msi::ColumnCategory::UpperCase.validate("Hello, World!"));

LowerCase

A text string containing no uppercase letters.

Examples

assert!(msi::ColumnCategory::LowerCase.validate("hello, world!"));
assert!(!msi::ColumnCategory::LowerCase.validate("Hello, World!"));

Integer

A signed 16-bit integer.

Examples

assert!(msi::ColumnCategory::Integer.validate("32767"));
assert!(msi::ColumnCategory::Integer.validate("-47"));
assert!(!msi::ColumnCategory::Integer.validate("40000"));

DoubleInteger

A signed 32-bit integer.

Examples

assert!(msi::ColumnCategory::DoubleInteger.validate("2147483647"));
assert!(msi::ColumnCategory::DoubleInteger.validate("-99999"));
assert!(!msi::ColumnCategory::DoubleInteger.validate("3000000000"));

Identifier

A string identifier (such as a table or column name). May only contain alphanumerics, underscores, and periods, and must start with a letter or underscore.

Examples

assert!(msi::ColumnCategory::Identifier.validate("HelloWorld"));
assert!(msi::ColumnCategory::Identifier.validate("_99.Bottles"));
assert!(!msi::ColumnCategory::Identifier.validate("3.14159"));

Property

A string that is either an identifier (see above), or a reference to an environment variable (which consists of a % character followed by an identifier).

Examples

assert!(msi::ColumnCategory::Property.validate("HelloWorld"));
assert!(msi::ColumnCategory::Property.validate("%HelloWorld"));
assert!(!msi::ColumnCategory::Property.validate("Hello%World"));

Filename

The name of a file or directory.

WildCardFilename

A filename that can contain shell glob wildcards.

Path

A string containing an absolute filepath.

Paths

A string containing a semicolon-separated list of absolute filepaths.

AnyPath

A string containing an absolute or relative filepath.

DefaultDir

A string containing either a filename or an identifier.

RegPath

A string containing a registry path.

Formatted

A string containing special formatting escapes, such as environment variables.

KeyFormatted

Unknown.

Template

Like Formatted, but allows additional escapes.

Condition

A string represeting a boolean predicate.

Guid

A hyphenated, uppercase GUID string, enclosed in curly braces.

Version

A string containing a version number. The string must consist of at most four period-separated numbers, with each number being at most 65535.

Examples

assert!(msi::ColumnCategory::Version.validate("1"));
assert!(msi::ColumnCategory::Version.validate("1.22"));
assert!(msi::ColumnCategory::Version.validate("1.22.3"));
assert!(msi::ColumnCategory::Version.validate("1.22.3.444"));
assert!(!msi::ColumnCategory::Version.validate("1.99999"));
assert!(!msi::ColumnCategory::Version.validate(".12"));
assert!(!msi::ColumnCategory::Version.validate("1.2.3.4.5"));

Language

A string containing a comma-separated list of deciaml language ID numbers.

Binary

A string that refers to a binary data stream.

CustomSource

A string that refers to a custom source.

Cabinet

A string that refers to a cabinet.

Shortcut

A string that refers to a shortcut.

Url

A string containing a URL.

Methods

impl ColumnCategory

fn validate(&self, string: &str) -> bool

Returns true if the given string is valid to store in a database column with this category.