Struct msi::Insert

pub struct Insert { /* fields omitted */ }

A database query to insert new rows.

Methods

impl Insert

fn into(table_name: &str) -> Insert

Starts building a query that will insert rows into the specified table.

fn row(self, values: Vec<Value>) -> Insert

Adds a new row to be inserted into the table.

fn rows(self, rows: Vec<Vec<Value>>) -> Insert

Adds multiple new rows to be inserted into the table.