Struct msi::Select

pub struct Select { /* fields omitted */ }

A database query to select rows.

Methods

impl Select

fn table(table_name: &str) -> Select

Starts building a query that will select rows from the specified table.

fn inner_join(self, rhs: Select, on: Expr) -> Select

Performs an inner join between this and another query, producing a row for each pair of rows from the two tables that matches the expression.

fn left_join(self, rhs: Select, on: Expr) -> Select

Performs a left join between this and another query.

fn columns(self, column_names: &[&str]) -> Select

Transforms the selected rows to only include the specified columns, in the order given.

fn with(self, condition: Expr) -> Select

Adds a restriction on which rows should be selected by the query; only rows that match the given boolean expression will be returned. (This method would have been called where(), to better match SQL, but where is a reserved word in Rust.)