Struct msi::SummaryInfo

pub struct SummaryInfo { /* fields omitted */ }

Summary information (e.g. title, author) about an MSI package.

Methods

impl SummaryInfo

fn author(&self) -> Option<&str>

Gets the "author" property, if one is set. This indicates the name of the person or company that created the package.

fn set_author(&mut self, author: String)

Sets the "author" property.

fn codepage(&self) -> CodePage

Gets the code page used for serializing this summary info.

fn set_codepage(&mut self, codepage: CodePage)

Sets the code page used for serializing this summary info.

fn comments(&self) -> Option<&str>

Gets the "comments" property, if one is set. This typically gives a brief description of the application/software that will be installed by the package.

fn set_comments(&mut self, comments: String)

Sets the "comments" property.

fn creating_application(&self) -> Option<&str>

Gets the "creating application" property, if one is set. This indicates the name of the software application/tool that was used to create the package.

fn set_creating_application(&mut self, app_name: String)

Sets the "creating application" property.

fn creation_time(&self) -> Option<SystemTime>

Gets the "creation time" property, if one is set. This indicates the date/time when the package was created.

fn set_creation_time(&mut self, timestamp: SystemTime)

Sets the "creation time" property.

fn set_creation_time_to_now(&mut self)

Sets the "creation time" property to the current time.

fn subject(&self) -> Option<&str>

Gets the "subject" property, if one is set. This typically indicates the name of the application/software that will be installed by the package.

fn set_subject(&mut self, subject: String)

Sets the "subject" property.

fn title(&self) -> Option<&str>

Gets the "title" property, if one is set. This indicates the type of the installer package (e.g. "Installation Database" or "Patch").

fn set_title(&mut self, title: String)

Sets the "title" property.

fn uuid(&self) -> Option<Uuid>

Gets the "UUID" property, if one is set.

fn set_uuid(&mut self, uuid: Uuid)

Sets the "UUID" property.