Struct msi::Table

pub struct Table { /* fields omitted */ }

A database table.

Methods

impl Table

fn name(&self) -> &str

Returns the name of the table.

fn columns(&self) -> &[Column]

Returns the list of columns in this table.

fn has_column(&self, column_name: &str) -> bool

Returns true if this table has a column with the given name.

fn get_column(&self, column_name: &str) -> Option<&Column>

Returns the column with the given name, if any.

fn primary_key_indices(&self) -> Vec<usize>

Returns the indices of table's primary key columns.